feat(events): bulk email check-in and fractional points
Points move from INTEGER to REAL so an award can be worth 0.75. Only
event_point_awards and retention_records change; crs_events.points is the
deprecated column and keeps its type rather than rebuilding a table with ten
incoming foreign keys.

Migration 0006 drops the PRAGMA foreign_keys lines drizzle-kit emits: D1
rejects them (7403) and the local runner wraps each file in a transaction
where they would no-op. Neither rebuilt table has an incoming FK, so the
rebuild is safe without them.

Input is quantized to 2dp at the zod boundary, so stored values stay clean and
sums never surface float dust. Retention thresholds and milestone targets stay
integer - they are policy numbers compared against a fractional total.

Bulk check-in reuses recordScan per row, so the window rules, point award,
dedupe and audit log are identical to a scan. Only the member lookup differs.
Unmatched emails come back to the officer as a "not found" list. The pasted
list is limited to event admins; a staffed scanner still checks members in one
at a time.

+++ src/db/contract/retention.ts
@@ -9,7 +9,7 @@ export const retentionRecordOutputSchema = z.object({
 	eventId: z.string().nullable(),
 	pointTypeId: z.string(),
 	pointTypeLabel: z.string(),
-	points: z.number().int().nullable(),
+	points: z.number().nullable(),
 	reason: z.string(),
 	source: z.enum(["event_attendance", "manual"]),
 	recordedBy: z.string(),
@@ -17,7 +17,7 @@ export const retentionRecordOutputSchema = z.object({
 });
 
 export const retentionSummaryOutputSchema = z.object({
-	totalPoints: z.number().int(),
+	totalPoints: z.number(),
 	recordCount: z.number().int(),
 	retainedAt: z.number().int(),
 	probationBelow: z.number().int(),
@@ -28,13 +28,13 @@ export const leaderboardRowOutputSchema = z.object({
 	memberId: z.string(),
 	fullName: z.string().nullable(),
 	name: z.string().nullable(),
-	totalPoints: z.number().int(),
+	totalPoints: z.number(),
 });
 
 export const myHistorySummaryOutputSchema = z.object({
 	termId: z.string(),
 	termName: z.string(),
-	totalPoints: z.number().int(),
+	totalPoints: z.number(),
 	retainedAt: z.number().int(),
 	probationBelow: z.number().int(),
 	status: z.enum(["retained", "on_track", "probation"]),
@@ -56,7 +56,7 @@ export const termMasterRowOutputSchema = z.object({
 	eventTitle: z.string().nullable(),
 	pointTypeId: z.string(),
 	pointTypeLabel: z.string(),
-	points: z.number().int().nullable(),
+	points: z.number().nullable(),
 	reason: z.string(),
 	source: z.enum(["event_attendance", "manual"]),
 	recordedAt: z.coerce.date(),
